refactor: rename app from AISightNew to AISight and simplify App.tsx structure
- Updated app name in app.json and package.json
- Simplified App.tsx by removing unnecessary imports and code
- Ensured proper navigation structure is maintained
- Fixed configuration for proper device deployment

This change aligns with the project created using:
npx @react-native-community/cli@latest init AISight --version 0.76.8
for safe development purposes.
