+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "AISightNew",
-  "displayName": "AISightNew"
+  "name": "AISight",
+  "displayName": "AISight"
 }
