+++ ios/AISight/Images.xcassets/AppIcon.appiconset/Contents.json
