+++ ios/AISight/PrivacyInfo.xcprivacy
