+++ ios/Podfile
@@ -14,7 +14,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'AISightNew' do
+target 'AISight' do
   config = use_native_modules!
 
   use_react_native!(
@@ -22,6 +22,11 @@ target 'AISightNew' do
     # An absolute path to your application root.
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
+
+  target 'AISightTests' do
+    inherit! :complete
+    # Pods for testing
+  end
 
   post_install do |installer|
     # https://github.com/facebook/react-native/blob/main/packages/react-native/scripts/react_native_pods.rb#L197-L202
