+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "AISightNew",
+  "name": "AISight",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -10,46 +10,35 @@
     "test": "jest"
   },
   "dependencies": {
-    "@maplibre/maplibre-react-native": "^10.2.1",
-    "@react-native-async-storage/async-storage": "^2.2.0",
-    "@react-native/new-app-screen": "0.81.4",
-    "@react-navigation/bottom-tabs": "^7.4.7",
     "@react-navigation/native": "^7.1.17",
     "@react-navigation/stack": "^7.4.8",
-    "axios": "^1.12.2",
-    "date-fns": "^4.1.0",
-    "react": "19.1.0",
-    "react-native": "0.81.4",
-    "react-native-bootsplash": "^6.3.11",
-    "react-native-config": "^1.5.9",
+    "react": "18.3.1",
+    "react-native": "0.76.8",
     "react-native-gesture-handler": "^2.28.0",
-    "react-native-reanimated": "^4.1.2",
     "react-native-safe-area-context": "^5.6.1",
-    "react-native-screens": "^4.16.0",
-    "react-native-vector-icons": "^10.3.0",
-    "zustand": "^5.0.8"
+    "react-native-screens": "^4.16.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.3",
     "@babel/runtime": "^7.25.0",
-    "@react-native-community/cli": "20.0.0",
-    "@react-native-community/cli-platform-android": "20.0.0",
-    "@react-native-community/cli-platform-ios": "20.0.0",
-    "@react-native/babel-preset": "0.81.4",
-    "@react-native/eslint-config": "0.81.4",
-    "@react-native/metro-config": "0.81.4",
-    "@react-native/typescript-config": "0.81.4",
-    "@types/jest": "^29.5.13",
-    "@types/react": "^19.1.0",
-    "@types/react-test-renderer": "^19.1.0",
+    "@react-native-community/cli": "15.0.1",
+    "@react-native-community/cli-platform-android": "15.0.1",
+    "@react-native-community/cli-platform-ios": "15.0.1",
+    "@react-native/babel-preset": "0.76.8",
+    "@react-native/eslint-config": "0.76.8",
+    "@react-native/metro-config": "0.76.8",
+    "@react-native/typescript-config": "0.76.8",
+    "@types/react": "^18.2.6",
+    "@types/react-test-renderer": "^18.0.0",
+    "babel-jest": "^29.6.3",
     "eslint": "^8.19.0",
     "jest": "^29.6.3",
     "prettier": "2.8.8",
-    "react-test-renderer": "19.1.0",
-    "typescript": "^5.8.3"
+    "react-test-renderer": "18.3.1",
+    "typescript": "5.0.4"
   },
   "engines": {
-    "node": ">=20"
+    "node": ">=18"
   }
 }
